+++ webpack.config.js
@@ -33,9 +33,14 @@ const config = {
             exclude: /node_modules/,
             use: [{
                 loader: 'ts-loader',
+                options: {
+                    compilerOptions: {
+                        "module": "es6" // override `tsconfig.json` so that TypeScript emits native JavaScript modules.
+                    }
+                }
             }]
         }]
     },
-};
+}
 
 module.exports = config;