+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.3
+- Add [truefire](https://marketplace.visualstudio.com/items?itemName=truefire.lilypond)'s VSCode extension as dependency
+- Update documentation and GIFs
+- Fix [#5](https://github.com/lhl2617/VSLilyPond/issues/5), revert to non-webpack version.
+
+
 ## 1.0.2
 - Minor package.json changes